fix: domain correctness, performance, API consistency (BD-1203)

+++ src/backend/Services/Documentation/BauDoku.Documentation.Application/Commands/DocumentInstallation/DocumentInstallationCommand.cs
@@ -18,7 +18,7 @@ public sealed record DocumentInstallationCommand(
     double? CorrectionAge,
     string? Description,
     string? CableType,
-    int? CrossSection,
+    decimal? CrossSection,
     string? CableColor,
     int? ConductorCount,
     int? DepthMm,
